+++ server/src/main/java/org/apache/druid/server/coordinator/compact/CompactionSegmentIterator.java
@@ -17,9 +17,8 @@
  * under the License.
  */
 
-package org.apache.druid.server.coordinator.duty;
+package org.apache.druid.server.coordinator.compact;
 
-import org.apache.druid.server.coordinator.CompactionStatistics;
 import org.apache.druid.timeline.DataSegment;
 
 import java.util.Iterator;
