+++ server/src/main/java/org/apache/druid/server/coordinator/compact/NewestSegmentFirstPolicy.java
@@ -17,8 +17,10 @@
  * under the License.
  */
 
-package org.apache.druid.server.coordinator.duty;
+package org.apache.druid.server.coordinator.compact;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.google.inject.Inject;
 import org.apache.druid.server.coordinator.DataSourceCompactionConfig;
 import org.apache.druid.timeline.SegmentTimeline;
 import org.joda.time.Interval;
@@ -27,16 +29,25 @@ import java.util.List;
 import java.util.Map;
 
 /**
- * Segment searching policy used by {@link CompactSegments}.
+ * This policy searches segments for compaction from the newest one to oldest one.
  */
-public interface CompactionSegmentSearchPolicy
+public class NewestSegmentFirstPolicy implements CompactionSegmentSearchPolicy
 {
-  /**
-   * Reset the current states of this policy. This method should be called whenever iterating starts.
-   */
-  CompactionSegmentIterator reset(
+  private final ObjectMapper objectMapper;
+
+  @Inject
+  public NewestSegmentFirstPolicy(ObjectMapper objectMapper)
+  {
+    this.objectMapper = objectMapper;
+  }
+
+  @Override
+  public CompactionSegmentIterator reset(
       Map<String, DataSourceCompactionConfig> compactionConfigs,
       Map<String, SegmentTimeline> dataSources,
       Map<String, List<Interval>> skipIntervals
-  );
+  )
+  {
+    return new NewestSegmentFirstIterator(objectMapper, compactionConfigs, dataSources, skipIntervals);
+  }
 }
